Stage to prod

+++ Subless.Models/MontlyPaymentStats.cs
@@ -5,7 +5,7 @@ namespace Subless.Models
     public class MontlyPaymentStats
     {
         public DateTimeOffset MonthStartDay { get; set; }
-        public int DollarsPaid { get; set; }
+        public double DollarsPaid { get; set; }
         public int Payers { get; set; }
     }
 }
